+++ back/controllers/messageCTRL.js
@@ -22,25 +22,26 @@ exports.getAll = (req, res, next) => {
 
   models.Message.findAll({
 
-    attributes: ['content','createdAt', 'id'],
+    attributes: ['id', 'content', "createdAt"],
     order: [
-        ['createdAt', 'ASC'],
-        [models.Comment, 'createdAt', 'ASC'],
+        ['createdAt', 'DESC'],
+        [models.Comment, 'createdAt', 'DESC'],
     ],
     include: [
-      {
-         model: models.User, as: 'user',
-         attributes: ['pseudo', 'id']
-      },
-      {
-        model : models.Comment,
-        attributes: ['content', 'createdAt'],
-        include: [ 
-          { model : models.User, as: 'user',
-          attributes: ['pseudo', 'id']
+        {
+            model: models.User,
+            attributes: ['pseudo', 'id']
+        },
+        {
+            model: models.Comment,
+            attributes: ['content', 'id', 'createdAt'],
+            include: [
+                {
+                    model: models.User,
+                    attributes: ['pseudo', 'id']
+                }
+            ]
         }
-        ]
-      }
     ]
   })
   .then((messages) => {
@@ -54,26 +55,58 @@ exports.getAll = (req, res, next) => {
 
 exports.deleteMessage = (req, res, next) => {
 
-  models.Message.destroy({
-      where: { id: req.params.id }
-  }) .then(() => {
-    res.status(201).json({ message : 'message suprimé'});
-  })
-  .catch((error) => {
-    res.status(400).json({ error });
-  })
+  try {
+    const userId = utilsToken.getId(req)
+    const userRole = utilsToken.getRole(req)
+
+    if ( userRole == "admin" || (req.body.userId == userId)) {
+        
+      models.Message.destroy({
+           where: { id: req.params.id }
+            }) .then(() => {
+      res.status(201).json({ message : 'message suprimé'});
+
+      })
+      .catch((error) => {
+        res.status(400).json({ error });
+      })
+
+    } else {
+        throw new Error(`Vous n'êtes pas autorisé(e) à réaliser cette action`);
+    }
+  } catch {
+    res.status(401).json({
+      error: new Error('Invalid request!')
+    });
+  }
+
 
 };
 
 exports.modifyMessage = (req, res, next) => {
+  try {
+    const userId = utilsToken.getId(req)
+    const userRole = utilsToken.getRole(req)
+    
+    if ( userRole == "admin" || (req.params.userId == userId)) {
 
-  const newContent = req.body.content
+        const newContent = req.body.content
+        
+        models.Message.update({ content: newContent },{ where: { id: req.params.id }})
+        .then(response => {
+         
+          if (response > 0) { res.status(200).json({ message: 'Méssage modifié' });
+          } else { res.status(400).json({ error: "Ce message n'existe pas"});
+        }
 
-  Message.update({ content: newContent },{ where: { id: req.params.id }})
-  .then(response => {
-    if (response > 0) { res.status(200).json({ message: 'Méssage modifié' });
-    } else { res.status(400).json({ error: "Ce message n'existe pas"});
+      })
+      .catch(error => res.status(500).json({ error}))
+    } else {
+      throw new Error(`Vous n'êtes pas autorisé(e) à réaliser cette action`);
   }
-})
-.catch(error => res.status(500).json({ error}))
+  } catch {
+    res.status(401).json({
+      error: new Error('Invalid request!')
+    });
+  }
 }
